+++ R/bage_mod-constructors.R
@@ -17,13 +17,12 @@
 #'
 #' - the name of a variable in `data`, with or without
 #'   quote marks, eg `"population"` or `population`;
-#' - the number `1`, in which case a pure "counts" model
+#' - `NULL`, in which case a pure "counts" model
 #'   with no exposure, is produced; or
 #' - `r lifecycle::badge("deprecated")`
-#'   a formula, which is evaluated with `data` as its
-#'   environment (see below for example). This option
-#'   has been deprecated, because it makes forecasting
-#'   and measurement error models more complicated.
+#'   the number `1`, in which case a pure "counts" model
+#'   is also produced (though this option is deprecated,
+#'   and will eventially be removed).
 #' 
 #' @section Mathematical details:
 #'
@@ -80,8 +79,8 @@
 #' specifying the outcome and predictors.
 #' @param data A data frame containing outcome,
 #' predictor, and, optionally, exposure variables.
-#' @param exposure Name of the exposure variable,
-#' or a `1`, or a formula. See below for details.
+#' @param exposure Name of the exposure variable
+#' or `NULL`. See below for details.
 #'
 #' @returns An object of class `bage_mod_pois`.
 #'
@@ -101,27 +100,15 @@
 #'   Detailed description of models
 #'
 #' @examples
-#' ## specify a model with exposure
+#' ## model with exposure
 #' mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
 #'                 data = nzl_injuries,
 #'                 exposure = popn)
 #'
-#' ## specify a model without exposure
+#' ## model without exposure
 #' mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
 #'                 data = nzl_injuries,
-#'                 exposure = 1)
-#'
-#' ## use a formula to specify exposure
-#' mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
-#'                 data = nzl_injuries,
-#'                 exposure = ~ pmax(popn, 1))
-#' ## but formulas are now deprecrated, and the
-#' ## recommended approach is to transform
-#' ## the input data outside the model:
-#' nzl_injuries$popn1 <- pmax(nzl_injuries$popn, 1)
-#' mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
-#'                 data = nzl_injuries,
-#'                 exposure = popn1)
+#'                 exposure = NULL)
 #' @export
 mod_pois <- function(formula,
                      data,
@@ -140,11 +127,15 @@ mod_pois <- function(formula,
   ## process 'exposure'
   if (!methods::hasArg(exposure))
     cli::cli_abort("Argument {.arg exposure} is missing, with no default.")
-  exposure <- deparse1(substitute(exposure))
-  exposure <- gsub("^\\\"|\\\"$", "", exposure)
-  is_offset_specified <- !identical(exposure, "1")
-  nm_offset_data <- if (is_offset_specified) exposure else NULL
+  nm_exposure <- deparse1(substitute(exposure))
+  if (identical(nm_exposure, "NULL"))
+    is_offset_specified <- FALSE
+  else {
+    nm_exposure <- gsub("^\\\"|\\\"$", "", nm_exposure)
+    is_offset_specified <- !identical(nm_exposure, "1")
+  }
   if (is_offset_specified) {
+    nm_offset_data <- nm_exposure
     check_offset_formula_not_used(nm_offset_data)
     check_offset_in_data(nm_offset_data = nm_offset_data,
                          nm_offset_mod = "exposure",
@@ -163,8 +154,10 @@ mod_pois <- function(formula,
     offset <- make_offset(nm_offset_data = nm_offset_data,
                           data = data)
   }
-  else
+  else {
+    nm_offset_data <- NULL
     offset <- make_offset_ones(data)
+  }
   ## check for suspicious rates
   if (is_outcome_in_data) {
     mult_high_rate <- 1000
@@ -194,18 +187,6 @@ mod_pois <- function(formula,
 #' which depend on the type of term (eg an intercept, an age main effect,
 #' or an age-time interaction.)
 #'
-#' @section Specifying size:
-#'
-#' The `size` argument can take two forms:
-#'
-#' - the name of a variable in `data`, with or without
-#'   quote marks, eg `"population"` or `population`; or
-#' - `r lifecycle::badge("deprecated")`
-#'   a formula, which is evaluated with `data` as its
-#'   environment (see below for example). This option
-#'   has been deprecated, because it makes forecasting
-#'   and measurement error models more complicated.
-#' 
 #' @section Mathematical details:
 #'
 #' The likelihood is
@@ -258,7 +239,7 @@ mod_pois <- function(formula,
 #' @param data A data frame containing the outcome
 #' and predictor variables, and the number of trials.
 #' @param size Name of the variable giving
-#' the number of trials, or a formula.
+#' the number of trials (with or without quote marks.)
 #'
 #' @returns An object of class `bage_mod`.
 #'
@@ -280,18 +261,6 @@ mod_pois <- function(formula,
 #' @examples
 #' mod <- mod_binom(oneperson ~ age:region + age:year,
 #'                  data = nzl_households,
-#'                  size = total)
-#'
-#' ## use formula to specify size
-#' mod <- mod_binom(ncases ~ agegp + tobgp + alcgp,
-#'                  data = esoph,
-#'                  size = ~ ncases + ncontrols)
-#' ## but formulas are now deprecrated, and the
-#' ## recommended approach is to transform
-#' ## the input data outside the model:
-#' esoph$total <- esoph$ncases + esoph$ncontrols
-#' mod <- mod_binom(ncases ~ agegp + tobgp + alcgp,
-#'                  data = esoph,
 #'                  size = total)
 #' @export
 mod_binom <- function(formula, data, size) {
@@ -374,16 +343,17 @@ mod_binom <- function(formula, data, size) {
 #' 
 #' @section Specifying weights:
 #'
-#' The `weights` argument can take three forms:
+#' There are three options for creating an unweighted
+#' model:
 #'
-#' - the name of a variable in `data`, with or without
-#'   quote marks, eg `"wt"` or `wt`;
-#' - the number `1`, in which no weights are used; or
-#' - `r lifecycle::badge("deprecated")`
-#'   a formula, which is evaluated with `data` as its
-#'   environment (see below for example). This option
-#'   has been deprecated, because it makes forecasting
-#'   and measurement error models more complicated.
+#' - do not supply a value for the `weights` variable;
+#' - set `weights` equal to `NULL`; or
+#' - `r lifecycle::badge("deprecated")` set weights equal
+#'   to `1`, though this option is deprecated, and will
+#'   eventually be removed.
+#'
+#' To create a weighted model, supply the name of
+#' the weighting variable in `data`, quoted or unquoted.
 #'
 #' @section Mathematical details:
 #'
@@ -458,23 +428,17 @@ mod_binom <- function(formula, data, size) {
 #'   Detailed description of models
 #'
 #' @examples
+#' ## model without weights
 #' mod <- mod_norm(value ~ diag:age + year,
-#'                 data = nld_expenditure,
-#'                 weights = 1)
+#'                 data = nld_expenditure)
 #'
-#' ## use formula to specify weights
-#' mod <- mod_norm(value ~ diag:age + year,
-#'                 data = nld_expenditure,
-#'                 weights = ~sqrt(value))
-#' ## but formulas are now deprecrated, and the
-#' ## recommended approach is to transform
-#' ## the input data outside the model:
+#' ## model with weights
 #' nld_expenditure$wt <- sqrt(nld_expenditure$value)
 #' mod <- mod_norm(value ~ diag:age + year,
 #'                 data = nld_expenditure,
 #'                 weights = wt)
 #' @export
-mod_norm <- function(formula, data, weights) {
+mod_norm <- function(formula, data, weights = NULL) {
   ## processing common to all models
   args <- mod_helper(formula = formula,
                      data = data,
@@ -482,13 +446,15 @@ mod_norm <- function(formula, data, weights) {
   outcome <- args$outcome
   is_outcome_in_data <- !is.null(outcome)
   ## process 'weights'
-  if (!methods::hasArg(weights))
-    cli::cli_abort("Argument {.arg weights} is missing, with no default.")
-  weights <- deparse1(substitute(weights))
-  weights <- gsub("^\\\"|\\\"$", "", weights)
-  is_offset_specified <- !identical(weights, "1")
-  nm_offset_data <- if (is_offset_specified) weights else NULL
+  nm_weights <- deparse1(substitute(weights))
+  if (identical(nm_weights, "NULL"))
+    is_offset_specified <- FALSE
+  else {
+    nm_weights <- gsub("^\\\"|\\\"$", "", nm_weights)
+    is_offset_specified <- !identical(nm_weights, "1")
+  }
   if (is_offset_specified) {
+    nm_offset_data <- nm_weights
     check_offset_formula_not_used(nm_offset_data)
     check_offset_in_data(nm_offset_data = nm_offset_data,
                          nm_offset_mod = "weights",
@@ -502,8 +468,10 @@ mod_norm <- function(formula, data, weights) {
     offset <- make_offset(nm_offset_data = nm_offset_data,
                           data = data)
   }
-  else
+  else {
+    nm_offset_data <- NULL
     offset <- make_offset_ones(data)
+  }
   offset_mean <- mean(offset, na.rm = TRUE)
   offset <- offset / offset_mean
   ## process outcome