+++ tests/testthat/test-check-functions.R
@@ -809,7 +809,7 @@ test_that("'check_mod_has_obs' returns correct error with no valid rows - no exp
     formula <- deaths ~ age + sex + time
     mod <- mod_pois(formula = formula,
                     data = data,
-                    exposure = 1)
+                    exposure = NULL)
     expect_error(check_mod_has_obs(mod),
                  "No data for fitting model.")
 })
@@ -1060,25 +1060,12 @@ test_that("'check_numeric' throws correct error with Inf", {
 test_that("'check_offset_formula_not_used' returns TRUE with valid offset", {
   expect_true(check_offset_formula_not_used(nm_offset_data = "popn"))
   rlang::local_options(lifecycle_verbosity = "warning")
-  expect_warning(check_offset_formula_not_used(nm_offset_data = "~popn + deaths"),
-                 "Using a formula to specify")
+  expect_error(check_offset_formula_not_used(nm_offset_data = "~popn + deaths"),
+               "Using a formula to specify")
 })
 
 
 ## 'check_offset_in_data' -----------------------------------------------------
-
-test_that("'check_offset_in_data' returns TRUE with valid formula", {
-    expect_true(check_offset_in_data(nm_offset_data = "~popn + deaths",
-                                     nm_offset_mod = "exposure",
-                                     data = data.frame(deaths = 1, popn = 2)))
-})
-
-test_that("'check_offset_in_data' throws correct error with invalid formula", {
-  expect_error(check_offset_in_data(nm_offset_data = "~popn + wrong",
-                                    nm_offset_mod = "exposure",
-                                    data = data.frame(deaths = 1, popn = 2)),
-               "Problem with formula used for `exposure`.")
-})
 
 test_that("'check_offset_in_data' returns TRUE with valid name", {
     expect_true(check_offset_in_data(nm_offset_data = "popn",
@@ -1102,15 +1089,7 @@ test_that("'check_offset_in_data' returns correct error with invalid name - size
 
 ## 'check_offset_nonneg' ----------------------------------------------------
 
-test_that("'check_offset_nonneg' returns TRUE with valid inputs - formula", {
-    expect_true(check_offset_nonneg(nm_offset_data = "~popn - deaths",
-                                    nm_offset_mod = "exposure",
-                                    data = data.frame(sex = 1:2,
-                                                      popn = c(0, 1.1),
-                                                      deaths = 0:1)))
-})
-
-test_that("'check_offset_nonneg' returns TRUE with valid inputs - name", {
+test_that("'check_offset_nonneg' returns TRUE with valid inputs", {
     expect_true(check_offset_nonneg(nm_offset_data = "popn",
                                     nm_offset_mod = "exposure",
                                     data = data.frame(sex = 1:2,
@@ -1118,16 +1097,7 @@ test_that("'check_offset_nonneg' returns TRUE with valid inputs - name", {
                                                       deaths = 0:1)))
 })
 
-test_that("'check_offset_nonneg' returns correct error with invalid inputs - formula", {
-  expect_error(check_offset_nonneg(nm_offset_data = "~popn - 1",
-                                   nm_offset_mod = "exposure",
-                                   data = data.frame(sex = 1:2,
-                                                     popn = c(-1, 1),
-                                                     deaths = 0:1)),
-               "`exposure` has negative value.")
-})
-
-test_that("'check_offset_nonneg' returns correct error with invalid inputs - name", {
+test_that("'check_offset_nonneg' returns correct error with invalid inputs", {
     expect_error(check_offset_nonneg(nm_offset_data = "popn",
                                     nm_offset_mod = "exposure",
                                     data = data.frame(sex = 1:3,
@@ -1188,6 +1158,10 @@ test_that("'check_old_version' raises error with invalid version", {
                "`object` appears to have been created with an old version of bage.")
   mod_has_stored <- mod
   mod_has_stored$seed_stored_draws <- 1
+  expect_error(check_old_version(mod_has_stored, nm_x = "object"),
+               "`object` appears to have been created with an old version of bage.")
+  mod_formula_offset <- mod
+  mod$nm_offset_data <- "~popn + 1"
   expect_error(check_old_version(mod_has_stored, nm_x = "object"),
                "`object` appears to have been created with an old version of bage.")
 })
@@ -1262,17 +1236,7 @@ test_that("'check_positive' throws correct error with valid inputs", {
 
 ## 'check_resp_le_offset' -----------------------------------------------------
 
-test_that("'check_resp_le_offset' returns TRUE with valid inputs - formula", {
-  data <- data.frame(deaths = c(0, 1, NA, 0,  NA),
-                     sex = rep("F", 5),
-                     popn =   c(0, 1, 2,  NA, NA))
-  expect_true(check_resp_le_offset(formula = deaths ~ sex,
-                                   nm_offset_data = "~  popn",
-                                   nm_offset_mod = "size",
-                                   data = data))
-})
-
-test_that("'check_resp_le_offset' returns TRUE with valid inputs - name", {
+test_that("'check_resp_le_offset' returns TRUE with valid inputs", {
   data <- data.frame(deaths = c(0, 1, NA, 0,  NA),
                      sex = rep("F", 5),
                      popn =   c(0, 1, 2,  NA, NA))
@@ -1282,18 +1246,7 @@ test_that("'check_resp_le_offset' returns TRUE with valid inputs - name", {
                                    data = data))
 })
 
-test_that("'check_resp_le_offset' raises correct error with invalid inputs - formula", {
-    data <- data.frame(deaths = c(0, 1, NA, 0,  2),
-                       sex = rep("F", 5),
-                       popn =   c(0, 1, 2,  NA, 1))
-    expect_error(check_resp_le_offset(formula = deaths ~ sex,
-                                      nm_offset_data = "~popn - 0.1",
-                                      nm_offset_mod = "size",
-                                      data = data),
-                 "Response greater than `size`.")
-})
-
-test_that("'check_resp_le_offset' raises correct error with invalid inputs - formula", {
+test_that("'check_resp_le_offset' raises correct error with invalid inputs", {
     data <- data.frame(deaths = c(0, 1, NA, 0,  2),
                        sex = rep("F", 5),
                        popn =   c(0, 1, 2,  NA, 1))
@@ -1307,17 +1260,7 @@ test_that("'check_resp_le_offset' raises correct error with invalid inputs - for
 
 ## 'check_resp_zero_if_offset_zero' -------------------------------------------
 
-test_that("'check_resp_zero_if_offset_zero' returns TRUE with valid inputs - formula", {
-    data <- data.frame(deaths = c(0, 1, NA, 0,  NA),
-                       sex = rep("F", 5),
-                       popn =   c(0, 1, 2,  NA, NA))
-    expect_true(check_resp_zero_if_offset_zero(formula = deaths ~ sex,
-                                               nm_offset_data = "~popn^2",
-                                               nm_offset_mod = "exposure",
-                                               data = data))
-})
-
-test_that("'check_resp_zero_if_offset_zero' returns TRUE with valid inputs - name", {
+test_that("'check_resp_zero_if_offset_zero' returns TRUE with valid inputs", {
     data <- data.frame(deaths = c(0, 1, NA, 0,  NA),
                        sex = rep("F", 5),
                        popn =   c(0, 1, 2,  NA, NA))
@@ -1327,18 +1270,7 @@ test_that("'check_resp_zero_if_offset_zero' returns TRUE with valid inputs - nam
                                                data = data))
 })
 
-test_that("'check_resp_zero_if_offset_zero' raises correct error with invalid inputs - formula", {
-  data <- data.frame(deaths = c(0, 1, NA, 0,  1),
-                     sex = rep("F", 5),
-                     popn =   c(0, 1, 2,  NA, 0))
-  expect_error(check_resp_zero_if_offset_zero(formula = deaths ~ sex,
-                                              nm_offset_data = "~popn^2",
-                                              nm_offset_mod = "exposure",
-                                              data = data),
-               "Response is non-zero but `exposure` is zero.")
-})
-
-test_that("'check_resp_zero_if_offset_zero' raises correct error with invalid inputs - name", {
+test_that("'check_resp_zero_if_offset_zero' raises correct error with invalid inputs", {
   data <- data.frame(deaths = c(0, 1, NA, 0,  1),
                      sex = rep("F", 5),
                      popn =   c(0, 1, 2,  NA, 0))